cleanup scripts

+++ sample-apps/error-processor/5-cleanup.sh
@@ -12,14 +12,18 @@ echo "Deleted $STACK stack."
 
 if [ -f bucket-name.txt ]; then
     ARTIFACT_BUCKET=$(cat bucket-name.txt)
-    while true; do
-        read -p "Delete deployment artifacts and bucket ($ARTIFACT_BUCKET)? (y/n)" response
-        case $response in
-            [Yy]* ) aws s3 rb --force s3://$ARTIFACT_BUCKET; rm bucket-name.txt; break;;
-            [Nn]* ) break;;
-            * ) echo "Response must start with y or n.";;
-        esac
-    done
+    if [[ ! $ARTIFACT_BUCKET =~ lambda-artifacts-[a-z0-9]{16} ]] ; then
+        echo "Bucket was not created by this application. Skipping."
+    else
+        while true; do
+            read -p "Delete deployment artifacts and bucket ($ARTIFACT_BUCKET)? (y/n)" response
+            case $response in
+                [Yy]* ) aws s3 rb --force s3://$ARTIFACT_BUCKET; rm bucket-name.txt; break;;
+                [Nn]* ) break;;
+                * ) echo "Response must start with y or n.";;
+            esac
+        done
+    fi
 fi
 
 while true; do
